ready for longterm hpc batches

+++ headers/QuartPolyEval.hpp
@@ -33,7 +33,7 @@ class QuartPolyEval{
         vector<double> imgSpaced;
         vector<vector<int>> binCountVect; // a large vect of rand ordered indices split into shorter vects
         vector<vector<int>> binCountVect2;
-		vector<vector<int>> binCountVect3;
+		vector<vector<unsigned short int>> binCountVect3;
 		vector<vector<int>> maxBinCountVect;
 		vector<vector<double>> pixelValVect; // a vect of vects to store absolute val of poly evaluated at each pixel
 		vector<vector<int>> colMinVect;
@@ -42,7 +42,7 @@ class QuartPolyEval{
 		vector<vector<int>> rowMaxVect;
 		vector<vector<int>> minPeaksVect;
 		vector<vector<vector<float>>> multPixValVect;
-		vector<vector<vector<int>>> multBinValVect;
+		vector<vector<vector<unsigned short int>>> multBinValVect;
 		float ***multPixValArr;
 		int ***multBinValArr;
 		int **binCountArr3;
@@ -77,7 +77,7 @@ class QuartPolyEval{
 		~QuartPolyEval() = default;
 		vector<vector<int>> getBinCount();
 		vector<vector<int>> getBinCount2();
-		vector<vector<int>> getBinCount3();
+		vector<vector<unsigned short int>> getBinCount3();
 		int **getBinCount4();
 		vector<vector<int>> getMaxBinCount();
 		vector<vector<double>> getPixelVal();
